REFACTOR: Removed useless index files
