+++ hooks/chat/use-stream-response.ts
@@ -1,6 +1,6 @@
 "use client";
 
-import { uploadAttachment } from "@/lib/actions/attachments";
+import { uploadAttachment } from "@/lib/actions/attachments/upload-attachment";
 import { persistMessage } from "@/lib/actions/chats/persist-message";
 import { reconstructThread } from "@/lib/chat/reconstruct-thread";
 import { useAppStore } from "@/lib/store";
