--- lib/actions/attachments/index.ts
@@ -1,7 +1,0 @@
-/**
- * Attachment-related server actions.
- *
- * @author Maruf Bepary
- */
-export * from "./get-attachment-url";
-export * from "./upload-attachment";
