--- lib/actions/prompts/index.ts
@@ -1,5 +1,0 @@
-export * from "./create-prompt";
-export * from "./list-prompts";
-export * from "./update-prompt";
-export * from "./delete-prompt";
-export * from "./get-prompt";
